convert/requirements.txt
pillow
requests
bytesbufio